+++ scrapypuppeteer/middleware/__init__.py
@@ -1,0 +1,3 @@
+from .service_middleware import PuppeteerServiceDownloaderMiddleware
+from .recaptcha_middleware import PuppeteerRecaptchaDownloaderMiddleware
+from .restore_middleware import PuppeteerContextRestoreDownloaderMiddleware
